+++ Application/DTOs/Payment/Vnpay/VnpayResponse.cs
@@ -22,7 +22,7 @@ namespace Application.DTOs.Payment.Vnpay
         public string vnp_TxnRef { get; set; } = string.Empty;
         public string vnp_SecureHashType { get; set; } = string.Empty;
         public string vnp_SecureHash { get; set; } = string.Empty;
-        public int? vnp_Amount { get; set; }
+        public decimal? vnp_Amount { get; set; }
         public string? vnp_ResponseCode { get; set; }
         public string vnp_PayDate { get; set; } = string.Empty;
